requirements.txt
neptune_client==0.11.0
nltk==3.6.5
numpy==1.21.4
pypesq==1.2.4
torch==1.9.1
torchtext==0.3.1
tqdm==4.62.2
